backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
asgiref==3.8.1
asknews==0.10.2
attrs==25.1.0
boto3==1.36.19
botocore==1.36.19
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
crontab==1.0.1
cryptography==43.0.3
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.8
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.10
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.18
langchain-aws==0.2.12
langchain-community==0.3.17
langchain-core==0.3.35
langchain-groq==0.2.4
langchain-mistralai==0.2.6
langchain-text-splitters==0.3.6
langgraph==0.2.72
langgraph-checkpoint==2.0.13
langgraph-sdk==0.1.51
langsmith==0.3.8
marshmallow==3.26.1
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.15
packaging==24.2
praw==7.8.1
prawcore==2.4.0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pypdf==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tenacity==9.0.0
tokenizers==0.21.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
yarl==1.18.3
zstandard==0.23.0
